requirements.txt
fastapi
uvicorn
torch==2.1.0+cpu
torchvision==0.16.0+cpu
numpy
opencv-python
pillow
python-multipart